Deep-Learning-Based-Temporal-Analysis-of-Badminton-Gameplay/requirements.txt
matplotlib==3.10.1
numpy==2.2.5
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
seaborn==0.13.2
streamlit==1.45.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
ultralytics==8.3.123
